Fix: Grammar & Structure
To me, this first line appears to be a general explanation of the app itself. So, I removed it as a top list item & changed it to the opening description. I rearranged the vertical lists so the introductory lines are now the list headings. Specifically, I removed the bullet point from the top list item & converted those phrases to title caps. Besides that, I substituted some phrases with ones that portray the feature more concisely; thus, improving readability. Another structural improvement includes the placement of spaces between the vertical lists. To interrupt the uniformity of the vertical list pattern (& keep your readers interested) I concluded the description with a sentence that conveys a few of the app’s notable features.